requirements.txt
django>=5.1.2
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-cors-headers
openai
python-dotenv